Update mep-feature-time-rail-thumbnails.js
If the plugin starts it bugs out a bit if the player has not started loading the video yet. The css height and width lines fixes the startup issue. I have also noticed that the image does not quite fit inside the block. This is due to the border pixels. (parseInt(w)+2) solves the border issue. Thanks a lot.